主键列均改为bigint，适应TPC-H SF500及更大数据量场景

+++ tpch/3.0.1/tpch-create-table.sql
@@ -3,21 +3,21 @@
 -- USE tpch;
 
 drop table if exists nation;
-create table nation  ( n_nationkey  integer not null,
+create table nation  ( n_nationkey  bigint not null,
 				n_name       char(25) not null,
-				n_regionkey  integer not null,
+				n_regionkey  bigint not null,
 				n_comment    varchar(152),
 				primary key(n_nationkey),
 				key nation_fk1 (n_regionkey) ) secondary_engine = rapid;
 
 drop table if exists region;
-create table region  ( r_regionkey  integer not null,
+create table region  ( r_regionkey  bigint not null,
 				r_name       char(25) not null,
 				r_comment    varchar(152),
 				primary key(r_regionkey) ) secondary_engine = rapid;
 
 drop table if exists part;
-create table part  ( p_partkey     integer not null,
+create table part  ( p_partkey     bigint not null,
 				p_name        varchar(55) not null,
 				p_mfgr        char(25) not null,
 				p_brand       char(10) not null,
@@ -29,10 +29,10 @@ create table part  ( p_partkey     integer not null,
 				primary key(p_partkey) ) secondary_engine = rapid;
 
 drop table if exists supplier;
-create table supplier ( s_suppkey     integer not null,
+create table supplier ( s_suppkey     bigint not null,
 				s_name        char(25) not null,
 				s_address     varchar(40) not null,
-				s_nationkey   integer not null,
+				s_nationkey   bigint not null,
 				s_phone       char(15) not null,
 				s_acctbal     decimal(15,2) not null,
 				s_comment     varchar(101) not null,
@@ -40,8 +40,8 @@ create table supplier ( s_suppkey     integer not null,
 				key supplier_fk1 (s_nationkey) ) secondary_engine = rapid;
 
 drop table if exists partsupp;
-create table partsupp ( ps_partkey     integer not null,
-				ps_suppkey     integer not null,
+create table partsupp ( ps_partkey     bigint not null,
+				ps_suppkey     bigint not null,
 				ps_availqty    integer not null,
 				ps_supplycost  decimal(15,2)  not null,
 				ps_comment     varchar(199) not null,
@@ -51,10 +51,10 @@ create table partsupp ( ps_partkey     integer not null,
 
 
 drop table if exists customer;
-create table customer ( c_custkey     integer not null,
+create table customer ( c_custkey     bigint not null,
 				c_name        varchar(25) not null,
 				c_address     varchar(40) not null,
-				c_nationkey   integer not null,
+				c_nationkey   bigint not null,
 				c_phone       char(15) not null,
 				c_acctbal     decimal(15,2)   not null,
 				c_mktsegment  char(10) not null,
@@ -63,8 +63,8 @@ create table customer ( c_custkey     integer not null,
 				key customer_fk1 (c_nationkey) ) secondary_engine = rapid;
 
 drop table if exists orders;
-create table orders  ( o_orderkey       integer not null,
-				o_custkey        integer not null,
+create table orders  ( o_orderkey       bigint not null,
+				o_custkey        bigint not null,
 				o_orderstatus    char(1) not null,
 				o_totalprice     decimal(15,2) not null,
 				o_orderdate      date not null,
@@ -76,10 +76,10 @@ create table orders  ( o_orderkey       integer not null,
 				key orders_fk1 (o_custkey) ) secondary_engine = rapid;
 
 drop table if exists lineitem;
-create table lineitem ( l_orderkey    integer not null,
-				l_partkey     integer not null,
-				l_suppkey     integer not null,
-				l_linenumber  integer not null,
+create table lineitem ( l_orderkey    bigint not null,
+				l_partkey     bigint not null,
+				l_suppkey     bigint not null,
+				l_linenumber  bigint not null,
 				l_quantity    decimal(15,2) not null,
 				l_extendedprice  decimal(15,2) not null,
 				l_discount    decimal(15,2) not null,
